Simple build pipeline using babel and webpack
This is to enable the use of the class system in the browswer, as otherwise require/import is not supported. Much more work
will need to be done on this branch to ensure it functions properly and can be used effectively, but it would be the better way
to organise the code and build.